feat: [] Next.js draft mode The guest space functionality is now passed through Next their draft mode. Users navigating to `/api/draft?secret=SECRET_TOKEN` enable the preview mode. A `slug` parameter can be added to be redirected to a specific page.
Additionally, the guest space can be enabled by passing `delivery_token`, `preview_token`, `space_id` and optionally a `domain` parameter. These parameters are handled in the `handleGuestSpaceParams` method that sets `previewData` on every request returned to `gSP` and `gSSP`. This data persists for as long as the draft mode is active.

`createPreviewClient` creates a GraphQL client, interpolated with the guest space parameters, and returns an API client with our codegenned `getSDK` method.

All pages conditionally use the production api client, or create a new preview client based on the `previewData` set from the `draft` api route.

Removed the `ctf-guest-space` feature. The React Query clientside hooks are no longer needed. React Query is still used for consent management, so the provider has been moved to the `ctf-consent-management` feature.

The `previewGraphQLClient` has been removed in favor of utility functions now living in the `pages/utils` folder. All preview functionality is now handled on the server, so the logic was moved there.